+++ tests/dummy/app/pods/components/custom-logo/component.js
@@ -1,0 +1,10 @@
+import Ember from 'ember'
+import layout from './template'
+
+const {
+  Component
+} = Ember
+
+export default Component.extend({
+  layout
+})