+++ tests/dummy/app/router.js
@@ -4,6 +4,7 @@ import config from './config/environment'
 var Router = Ember.Router.extend({
   location: config.locationType
 })
+// BEGIN-SNIPPET router
 Router.map(function () {
   this.nav('demo', {
     path: '/'
@@ -15,26 +16,27 @@ Router.map(function () {
       this.column('Column 1', {
         color: '#009eef'
       }, function () {
-        this.action('Action 1', {
+        this.action('Action Example', {
           action: 'myAction',
           pack: 'dummy',
           icon: 'sample',
-          description: 'My description',
+          description: 'This is an action',
           inline: true
         })
-        this.app('App 1', {
+        this.app('Route Example', {
           route: 'go',
-          description: 'Description 1',
+          description: 'This is a route',
           pack: 'dummy',
           icon: 'sample'
         })
         this.engine('Blog Engine', {
           route: 'blog',
+          description: 'This is an engine example',
           package: 'ember-blog-engine',
           pack: 'dummy',
           icon: 'sample'
         })
-        this.section('Section 1', {
+        this.section('More Content', {
           color: '#a1e7ff'
         }, function () {
           this.link('Google', {
@@ -52,9 +54,10 @@ Router.map(function () {
           })
         })
       })
-      this.column('Column 2')
+      this.column('Empty Column')
     })
   })
 })
 
 export default Router
+// END-SNIPPET